+++ RiskOfChaos/ModifierController/Effect/IEffectModificationProvider.cs
@@ -1,0 +1,6 @@
+﻿namespace RiskOfChaos.ModifierController.Effect
+{
+    public interface IEffectModificationProvider : IValueModificationProvider<EffectModificationInfo>
+    {
+    }
+}